requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.22
psycopg2-binary>=2.9.9
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dateutil>=2.8.2
requests>=2.31.0
kafka-python>=2.0.2
GitPython>=3.1.41
neo4j>=5.16.0
python-dotenv>=1.0.0